+++ app/src/main/java/com/ad430/globaltactics/EventsFragment.java
@@ -1,22 +1,20 @@
 package com.ad430.globaltactics;
 
 import android.app.Activity;
-import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toolbar;
+
+import com.ad430.globaltactics.adapters.EventAdapter;
+import com.ad430.globaltactics.objects.Event;
+import com.ad430.globaltactics.viewmodels.EventViewModel;
 
 import java.util.ArrayList;
 import java.util.Collections;
