+++ app/src/main/java/com/ad430/globaltactics/OurExpertsFragment.java
@@ -3,7 +3,6 @@ package com.ad430.globaltactics;
 import android.app.Activity;
 import android.os.Bundle;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -14,11 +13,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.QueryDocumentSnapshot;
-import com.google.firebase.firestore.QuerySnapshot;
+import com.ad430.globaltactics.adapters.OurExpertsAdapter;
+import com.ad430.globaltactics.objects.Expert;
+import com.ad430.globaltactics.viewmodels.OurExpertsViewModel;
 
 import java.util.ArrayList;
