+++ app/src/main/java/com/ad430/globaltactics/OurSuccessesCorporateTab.java
@@ -8,14 +8,16 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.ad430.globaltactics.adapters.EventAdapter;
+import com.ad430.globaltactics.adapters.OurSuccessTabListAdapter;
+import com.ad430.globaltactics.objects.OurSuccess;
+import com.ad430.globaltactics.viewmodels.OurSuccessesViewModel;
+
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 
 public class OurSuccessesCorporateTab extends Fragment {
