+++ app/src/main/java/com/ad430/globaltactics/objects/Expert.java
@@ -1,4 +1,4 @@
-package com.ad430.globaltactics;
+package com.ad430.globaltactics.objects;
 
 import android.util.Log;
